+++ lib/reportinator/config.rb
@@ -6,10 +6,24 @@ module Reportinator
     }
     DEFAULT_REPORT_DIRS = ["reports", "app/reports"]
     DEFAULT_REPORT_SUFFIXES = ["report.json", "json"]
+    DEFAULT_FUNCTIONS = [
+      "Reportinator::JoinArrayFunction",
+      "Reportinator::MethodArrayFunction",
+      "Reportinator::AdditionStringFunction",
+      "Reportinator::ConstantStringFunction",
+      "Reportinator::DateStringFunction",
+      "Reportinator::JoinStringFunction",
+      "Reportinator::LogicalStringFunction",
+      "Reportinator::NumberStringFunction",
+      "Reportinator::RangeStringFunction",
+      "Reportinator::SymbolStringFunction",
+      "Reportinator::VariableStringFunction"
+    ]
 
     attribute :report_directories, default: []
     attribute :report_suffixes, default: []
     attribute :report_types, default: {}
+    attribute :parser_functions, default: []
     attribute :output_directory, default: "reports"
 
     def configured_directories
@@ -23,6 +37,11 @@ module Reportinator
     def configured_types
       types = DEFAULT_TYPES
       types.merge(report_types)
+    end
+
+    def configured_functions
+      functions = DEFAULT_FUNCTIONS + parser_functions
+      functions.map { |function| function.constantize }
     end
   end
 end