+++ lib/reportinator/functions/array/method.rb
@@ -1,0 +1,9 @@
+module Reportinator
+  class MethodArrayFunction < ArrayFunction
+    PREFIXES = ["#"]
+
+    def output
+      parse_and_execute_value(target, values)
+    end
+  end
+end