+++ lib/reportinator/functions/string/constant.rb
@@ -1,0 +1,9 @@
+module Reportinator
+  class ConstantStringFunction < StringFunction
+    PREFIXES = ["&"]
+
+    def output
+      body.constantize
+    end
+  end
+end