+++ lib/reportinator/functions/string/date.rb
@@ -1,0 +1,9 @@
+module Reportinator
+  class DateStringFunction < StringFunction
+    PREFIXES = ["!d"]
+
+    def output
+      Time.parse(body)
+    end
+  end
+end