+++ lib/reportinator/functions/string/number.rb
@@ -1,0 +1,11 @@
+module Reportinator
+  class NumberStringFunction < StringFunction
+    PREFIXES = ["!n"]
+
+    def output
+      float = (body =~ /\d\.\d/)
+      return body.to_f if float.present?
+      body.to_i
+    end
+  end
+end