+++ lib/reportinator/functions/string/symbol.rb
@@ -1,0 +1,9 @@
+module Reportinator
+  class SymbolStringFunction < StringFunction
+    PREFIXES = [":"]
+
+    def output
+      body.to_sym
+    end
+  end
+end