+++ lib/reportinator/functions/string/variable.rb
@@ -1,0 +1,10 @@
+module Reportinator
+  class VariableStringFunction < StringFunction
+    PREFIXES = ["$"]
+
+    def output
+      return element unless variables.include? body.to_sym
+      variables[body.to_sym]
+    end
+  end
+end